requirements.txt
PyQt6>=6.4.0
SQLAlchemy>=2.0.0
pandas>=2.0.0
matplotlib>=3.7.0
pyface>=8.0.0
bcrypt>=4.0.0
pillow>=9.0.0
python-dotenv>=1.0.0
pytest>=7.0.0
pyobjc-framework-LocalAuthentication>=9.2
keyring>=23.0.0
auth >= 1.0.0